+++ packages/@vangware/iterables/src/getIterator.ts
@@ -1,4 +1,8 @@
-import { isIterable } from "@vangware/predicates";
+import {
+	asyncIteratorSymbol,
+	isIterable,
+	iteratorSymbol,
+} from "@vangware/predicates";
 import type { IsomorphicIterable } from "@vangware/types";
 import type { ReadOnlyAsyncIterable } from "./types/ReadOnlyAsyncIterable.js";
 import type { ReadOnlyAsyncIterator } from "./types/ReadOnlyAsyncIterator.js";
@@ -24,9 +28,9 @@ export const getIterator = <Iterable extends IsomorphicIterable>(
 	iterable: Iterable,
 ) =>
 	(iterable as AsyncIterable<unknown>)[
-		Symbol[
-			isIterable(iterable) ? "iterator" : "asyncIterator"
-		] as keyof AsyncIterable<unknown>
+		(isIterable(iterable)
+			? iteratorSymbol
+			: asyncIteratorSymbol) as keyof AsyncIterable<unknown>
 	]() as Iterable extends IsomorphicIterable<infer Item>
 		? Iterable extends ReadOnlyAsyncIterable<Item>
 			? ReadOnlyAsyncIterator<Item, Item, Item>