+++ packages/@vangware/predicates/src/isAsyncIterable.ts
@@ -2,6 +2,7 @@ import type { ReadOnly } from "@vangware/types";
 import { hasAsyncIteratorSymbol } from "./hasAsyncIteratorSymbol.js";
 import { isFunction } from "./isFunction.js";
 import { isObject } from "./isObject.js";
+import { asyncIteratorSymbol } from "./symbols.js";
 
 /**
  * Check if given value is `AsyncIterable`.
@@ -24,4 +25,4 @@ export const isAsyncIterable = <Item>(
 ): input is ReadOnly<AsyncIterable<Item>> =>
 	isObject(input) &&
 	hasAsyncIteratorSymbol(input) &&
-	isFunction(input[Symbol.asyncIterator]);
+	isFunction(input[asyncIteratorSymbol]);