+++ packages/preact-pair/src/pair.ts
@@ -1,5 +1,6 @@
-import type { Function, Unary } from "@vangware/types";
-import type { h } from "preact";
+import type { Function } from "@vangware/types";
+import { mutate } from "@vangware/utils";
+import type { FunctionComponent } from "preact";
 import type { PairedComponentProperties } from "./PairedComponentProperties.js";
 
 /**
@@ -40,12 +41,6 @@ import type { PairedComponentProperties } from "./PairedComponentProperties.js";
  * @returns Component that expects a function as children with the paired hook.
  */
 export const pair = <Hook extends Function>(hook: Hook) =>
-	Object.assign(
-		(({ children }) => children(hook)) as Unary<
-			PairedComponentProperties<Hook>,
-			ReturnType<typeof h>
-		>,
-		{ displayName: `paired(${hook.name})` },
-	) as Unary<PairedComponentProperties<Hook>, ReturnType<typeof h>> & {
-		readonly displayName: `paired(${string})`;
-	};
+	mutate({ displayName: `paired(${hook.name})` })(({ children }) =>
+		children(hook),
+	) as FunctionComponent<PairedComponentProperties<Hook>>;